cleanup: move fill_pressures from dive.c to gas.c
This function does not access a dive structure.
